added changes for user product page
