+++ countDownFrontEnd/src/app/books/books.component.css
@@ -41,7 +41,7 @@
 }
 
 .card-body {
-    height: 70px;
+    height: 10px;
     padding-bottom: 0px;
 }
