+++ countDownFrontEnd/src/app/dashboard/shared/product.model.ts
@@ -1,0 +1,6 @@
+export class Product {
+    Id: number;
+    Title: string;
+    ReleaseDate: string;
+    ImgLink: string;
+}
