+++ countDownFrontEnd/src/app/dashboard/user-products/user-product-list/user-product-list.component.html
@@ -1,0 +1,1 @@
+<p>user-product-list works!</p>
