+++ countDownFrontEnd/src/app/dashboard/user-products/user-product-list/user-product-list.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-user-product-list',
+  templateUrl: './user-product-list.component.html',
+  styles: []
+})
+export class UserProductListComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
