+++ countDownFrontEnd/src/app/dashboard/user-products/user-product/user-product.component.html
@@ -1,0 +1,16 @@
+<p>This is where the user custom-product form will go</p>
+<form #form="ngForm" autocomplete="off" (submit)="onSubmit(form)">
+   <input type="hidden" name="Id" [value]="service.formData.Id">
+    <div class="form-group"> 
+    <input name="Title" #Title="ngModel" [(ngModel)] = "service.Title" class="form-control" placeholder="Product's Title">
+    </div>
+    <div class="form-group"> 
+    <input name="ReleaseDate" #ReleaseDate="ngModel" [(ngModel)] = "service.ReleaseDate" class="form-control" placeholder="Release Date">
+    </div>
+    <div class="form-group"> 
+    <input name="ImgLink" #ImgLink="ngModel" [(ngModel)] = "service.ImgLink" class="form-control" placeholder="full link to an image">
+    </div>
+    <div class="form-group"> 
+    <button type="submit">Submit</button>
+    </div>
+</form>
