+++ countDownFrontEnd/src/app/dashboard/user-products/user-products.component.html
@@ -1,0 +1,3 @@
+<p>
+    user products works!
+</p>
