+++ countDownFrontEnd/src/app/dashboard/user-products/user-products.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-user-products',
+  templateUrl: './user-products.component.html',
+  styles: []
+})
+export class UserProductsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
