+++ countDownFrontEnd/src/app/movies/movies.component.html
@@ -1,1 +1,63 @@
-<p>movies works!</p>
+<h3>Movies to Watch</h3>
+<div class="card-container" >
+    <!-- LOGGED IN USER CARD WITH ADD AND SHARE BUTTONS USING INFO ICON TO RETRIEVE MORE EVENT INFORMATION -->
+    <!-- ######################################## SINGLE CARD 1 ########################################  -->
+    <!-- ######################################## SINGLE CARD 1 ########################################  -->
+    <!-- OVER ALL CARD STYLING  -->
+    <div class="card text-white mb-3" *ngFor ="let movie of movies">
+        <!-- CARD HEADER STARTS -->
+        <div id="card-header" class="card-header ml-auto">
+            <!-- CLOCK -->
+            <div class="count">
+                <div class="countd"><span id="day">{{daysUntilRelease(movie.release_date)}}</span>Days</div>
+                <div class="countd"><span id="hour">{{hoursUntilRelease(movie.release_date)}}</span>Hours</div>
+                <div class="countd"><span id="minute">{{minsUntilRelease(movie.release_date)}}</span>Minutes</div>
+                <div class="countd"><span id="second">{{secUntilRelease(movie.release_date)}}</span>Seconds</div>
+            </div>
+        </div>
+        <!-- CARD IMAGE SOMETHING WEIRD HAPPENS HERE  -->
+​
+        <img src="https://image.tmdb.org/t/p/w500{{movie.poster_path}}" class="card-img-top" alt="...">
+        <!-- CARD BODY STARTS  -->
+        <div class="card-body">
+            <!-- CARD TEXT STARTS  -->
+            <!-- <h5 class="card-title"><span> 
+                <img class="film-icon" src="/images/film.png" alt="">
+                {{movie.title}}</span></h5> -->
+        </div>
+        <!-- CARD BUTTONS AT BOTTOM  -->
+        <div class="card-buttons">
+            <button type="button" class="btn share-icon"><i class="fa fa-share" ></i></button>
+            <button type="button" class="btn info-icon"><i class="fa fa-info-circle"></i></button>
+            <button type="button" class="btn add-icon"><i class="fa fa-plus"></i></button>
+            <!-- <button type="button" class="btn add-icon"><i class="fa fa-plus" ng-click="PostData()></i></button> -->
+        </div>
+      </div>
+    </div>
+
+<!--Probably could make this script code into a getCountdown function inside an load event-->
+
+
+
+ <script>
+     let count = new Date("Nov 21, 2021 00:00:00").getTime();
+    let anothercount = new Date("June 5, 2021 00:00:00").getTime();
+    let x = setInterval(function(){
+        let now = new Date().getTime();
+        let d = count - now;
+
+        let day = Math.floor(d/(1000 * 60 * 60 * 24));
+        let hour = Math.floor((d % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60));
+        let minute = Math.floor((d % (1000 * 60 * 60)) / (1000 * 60));
+        let second = Math.floor((d % (1000 * 60)) / 1000);
+
+        document.getElementById('day').innerHTML = day;
+        document.getElementById('hour').innerHTML = hour;
+        document.getElementById('minute').innerHTML = minute;
+        document.getElementById('second').innerHTML = second;
+
+        if(d <= 0){
+            clearInterval(x);
+        }
+    }, 1000)
+ </script>
